parser/requirements.txt
python_dateutil==2.8.0
ijson==2.4
jsonschema==3.0.1
